Improve for public use.
- SearchRecipe is now to be configured in Cookbook.
- User must subclass SearchRecipe and override filter method.
