+++ source/SearchResult.xml
@@ -1,3 +1,8 @@
-<shelf><header><title>Results</title></header><section id="Results">
-{{results}}
-</section></shelf>
+<shelf>
+    <header>
+        <title>Results</title>
+    </header>
+    <section id="Results">
+        {{results}}
+    </section>
+</shelf>
